requirements.txt
streamlit
groq
sqlalchemy
pandas
plotly
pymysql
psycopg2-binary
python-dotenv
sqlparse
